Unified ProxyConfigList tests
